Fix freeing nodes with maxed refcounts
If a node reaches the maximum refcount, it is added to the `d_maxedOut`
vector in `NodeManager` and not destroyed as long as the `NodeManager`
lives. When the node manager is destroyed, the nodes in that vector are
sorted topologically (bottom-up) and we use the vector to destroy the
nodes top-down (i.e.  we destroy nodes representing subterms after we
destroy the terms containing them). Unfortunately, the sorting function
`NodeManager::TopologicalSort()` had a wrong assertion, which assumed
that a term could not be in the `visited` set if we were visiting it in
preorder. This assertion could be violated by having a node such as
`(and x x)` because `x` would be added to `stack` twice and the first
visit of `x` would add it to the `visited` set. The second visit would
then trigger the assertion. This PR removes the assertion and refactors
the code to make it easier to understand.