Refactor parser to define fewer tokens for symbols
For some function symbols (`*`, `<`, etc.), we used tokens instead of
handling them like other function symbols. As far as I can tell, this is
a leftover from a design change that occurred a long time ago but was
not applied to all existing function symbols. Having the function
symbols as tokens is worse because you have to be careful to disable
them when a theory is not in use and we were not doing this properly in
multiple instances (see issue #2918), so we couldn't do for example
`(define-const * Bool)` even when no arithmetic theory was active. In
the current design, we define function symbols using
`Smt2::addOperator()` depending on the active theories.

This commit makes multiple changes related to that:
- It removes `builtinOp` from Smt2.g in favor of defining all operators
  in `Smt2` and moves the code that deals with some exceptional cases to
  the normal function application parsing code.
- For indexed operators, it introduces a similar mechanism via
  `addIndexedOperator()` and removes all the corresponding parsing cases
  from Smt2.g, replacing them with `mkIndexedOperator()` which takes as an
  input a name and the integer arguments to create an `OpTerm`
  corresponding to the indexed operator.
- For constants (e.g. `re.allchar`), it uses `defineVar()` to defines
  the expression that the consts correspond to.
- For indexed constants (e.g. `(_ +oo 8 24)`), it introduces a new
  method `Smt2::mkIndexedConstant()` that takes as an input the name of
  the constant and the integer parameters and returns a `Term`
  corresponding to the constant.
- Some of the regressions were broken by the stricter checking of
  defined symbols, e.g. multiplication would parse regardless of the
  logic.
- Several fixes to which operators are allowed in strict mode.

Overall, this commit moves logic out of the parser and uses `Term`s and
`OpTerm`s in some more places than before.