Fix concat-find regexp elimination
Fixes #2982. The `concat-find` regexp elimination has an off-by-one
error in the upper bound of the quantifier. For example, for `(str.in.re
x (re.++ R1 "A" R2))`, the upper bound for the quantifier variable in
the eliminated term was `(< k (- (str.len x) 1))`. However, it should be
`(< k (- (str.len x) 1))` because we have to account for the case where
`"A"` is at the end of the string, i.e. at position `(- (str.len x) 1)`.
This commit fixes the issue and moves two assertions that were too
aggressive.