Add run script for next SMT-COMP
This commit adds the run scripts for next SMT-COMP. They are just copies
of the 2019 files for now. Note that the new files do not mention a
year.  Instead, the idea is to keep working on these new files and then
just copy them after every year's competition. Like this, we have all
the changes over the years in one file instead of spread across multiple
files.