Remove casts to subclasses of Type in API
Casting to subclasses of `Type` is dangerous. The copy constructor and
the assignment operator ot `Type` both turn a subclass of `Type` (e.g.
`DatatypeType`) into an instance of `Type`. One way of this happening is
the constructor of `Sort` which takes a `Type` (e.g.
[this](https://github.com/CVC4/CVC4/blob/8dda9531995953c3cec094339002f2ee7cadae08/src/api/cvc4cpp.cpp#L2954)
conversion indirectly turns a `Datatype` into a `Type`). For the
sort-specific methods, e.g. `Sort::getFunctionDomainSorts()`, the API
was trying to cast the `Type` in `d_type` into the corresponding
subclass, e.g. `FunctionType`, but the `Sort` class was holding an
instance of `Type` and not `FunctionType`, so the conversion was invalid
and UBSan complained. The class hierarchy of `Type`s is somewhat unusual
in that none of the subclasses have additional members (they just offer
additional ways of accessing information of a given `Type` instance) and
the subclasses all support a constructor that converts a type into the
corresponding subclass. Note this is probably also why the conversion to
a wrong type did not cause observable issues. This commit fixes the
issue by converting `Type`s using the formerly mentioned constructors.