[Regressions] Support for running w/ default args
When we have regression tests with multiple sets of `COMMAND-LINE`
arguments, it is sometimes desribable to include a run with the default
arguments. However, our regression script was ignoring lines of the form
`COMMAND-LINE:` without a space after the colon, so the test was not run
with the default arguments. This commit changes the detection of the
regression commands to not require the space after the colon.