Fix reduction of `sqrt`
Fixes #3475. This commit fixes two issues related to the reduction of
`sqrt`. Previously, we reduced `sqrt(x)` to `choice y. y = x * x`. The
problem is that this reduction makes a problem `unsat` as soon as there
is a square root of a negative number and it does not maintain the
invariant that `x = y => sqrt(x) = sqrt(y)` (i.e. that `sqrt` behaves
like a function). This commit changes the reduction of `sqrt(x)` to
`choice y. ite(x >= 0.0, y * y = x ^ Uf(x), Uf(x))`, meaning that we
interpret the square root of negative numbers as an undefined value
(similar to divisions by zero in the theories of integer/real
arithmetic).