Bi-directional unrolling of R* regular expressions
This commit changes the unfolding of `x in R*`. Previously, for the case
where `x` matches two or more `R`s, we were unfolding it to `x = x1 ++
x2 ^ x1 != "" ^ x2 != "" ^ x1 in R ^ x2 in R*`. This made it difficult
to find conflicts with the end of `x`. The new unfolding, unfolds both
the beginning and the end: `x = x1 ++ x2 ++ x3 ^ x1 != "" ^ x3 != "" ^
x1 in R ^ x2 in R* ^ x3 in R)` (note that we do not require `x2` to be
non-empty). This allows us to reason about both the beginning and the
end of `x` simultaneously.