Fix printing of models of uninterpreted sorts
Fixes #3596. When printing the representatives of uninterpreted sorts in
a model, CVC4 would print lines such as:

```
(declare-fun  () integer)
```

This was due to two issues:

- Representatives of uninterpreted sorts were not quoted correctly
  (empty symbol names should be quoted)
- The `quoteSymbol` function in `smt2_printer.cpp` was not turning a
  node into a string before quoting it but was instead just generating
  an empty string

This commit fixes both issues and unifies the implementations for
quoting symbols that we had in `printer/smt2/smt2_printer.cpp` and
`util/smt2_quote_string.cpp`.