Better heuristics for marking congruent variables
The string solver marks all but one variable in an equivalence class as
congruent. This commit improves that heuristic to choose the oldest
variable as the one that does not get marked as congruent. Empirically
and intuitively this is a good idea because it is less likely that we
get into loops while reasoning.