Fix arithmetic rewriter for exponential
Fixes #3683. When rewriting `exp(2 + x)`, we did not reach a fix-point
after a single call to the rewriter. We rewrote it to `exp(2) * exp(x)`
instead of rewriting `exp(2)` further. This commit fixes the issue by
making the rewrite response `REWRITE_AGAIN_FULL` s.t.  we rewrite the
subterms of the term we created.