Make "unknown" non-critical for unsat cores check
Fixes #3480. Fixes #3655. `--check-unsat-cores` was failing when we got
`unknown` as an answer. However, for undecidable theories, we cannot
guarantee that we will be able to determine that the unsat core by
itself is `unsat`. This commit weakens the check to emit a warning
instead of an error when the answer is `unknown`.