Remove parser from bindings
This commit removes the parser from the bindings. Note that it is
already in a broken state currently (the reason why CVC4Streams and
PipedInput are removed from the Java API examples) and it does not make
sense to fix it for the old API. If we want to have some parser
facilities with the new API, we should think about a more minimal way of
exposing those pieces. This should also fix the bindings issue in #3767
and is work towards #2846.